Moving palette_map functions in submodule of color
